- Fixed: The new sound pausing on menu open interrupted Strife conversations. - Did some very preliminary support for x64 compiling. - Now I see why makewad is using the __fastcall calling convention by default:   Because the release zlib is built with it as well. - Fixed: Conversion from 'const char *' to 'char *' in DEM_SAVEGAME case in   Net_DoCommand(). - Updated the project files for Visual C++ 2005.
SVN r138 (trunk)

+++ tools/re2c/mbo_getopt.cc
@@ -42,7 +42,7 @@ int mbo_getopt(int argc, char* const *argv, const mbo_opt_struct opts[], char **
 {
 	static int optchr = 0;
 	static int dash = 0; /* have already seen the - */
-	int arg_start = 2;
+	size_t arg_start = 2;
 
 	int opts_idx = -1;
